% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock.R
\name{axl_rows}
\alias{axl_rows}
\title{Axl Rows}
\usage{
axl_rows(.data, ..., play = FALSE)
}
\arguments{
\item{.data}{A data frame to apply axl-row-wise operations on.}

\item{...}{Additional arguments passed to \code{\link[dplyr]{rowwise}()}.}

\item{play}{Logical, whether to play the track on Spotify.}
}
\description{
Axl-row-wise operations for your data frame.
}
