% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funk.R
\name{bruno_sparse}
\alias{bruno_sparse}
\title{Bruno sparse}
\usage{
bruno_sparse(.data, ..., play = FALSE)
}
\arguments{
\item{.data}{The dataframe of a vector to subsample from.}

\item{...}{Additional arguments passed to \code{slice_sample()} or \code{sample()}.}

\item{play}{Logical, whether to play the track on Spotify.}
}
\description{
Make your data Bruno sparse.
}
