% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock.R
\name{dir_straights}
\alias{dir_straights}
\title{Dir straights}
\usage{
dir_straights(dir, ...)
}
\arguments{
\item{dir}{A directory path to be normalized.}

\item{...}{Additional arguments passed to \code{\link[base]{normalizePath}()}.}
}
\description{
Straighten your directory path.
}
