% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock.R
\name{fill_collins}
\alias{fill_collins}
\title{Fill Collins}
\usage{
fill_collins(.data, ...)
}
\arguments{
\item{.data}{A data frame to be modified.}

\item{...}{Additional arguments to be passed to the mutate function.}
}
\description{
Fill a collin!
}
