% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{liszt.files}
\alias{liszt.files}
\title{Liszt your files}
\usage{
liszt.files(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[base]{list.files}()}.}
}
\description{
Liszt all files in a directory.
}
