% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicplyr-package.R
\docType{package}
\name{musicplyr-package}
\alias{musicplyr}
\alias{musicplyr-package}
\title{musicplyr: Play music in your head while programming}
\description{
Play music in your head while programming
}
\author{
\strong{Maintainer}: Tobias Hundertmark \email{Tobias.100mark@outlook.de}

}
\keyword{internal}
