% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetys.R
\name{n_sync}
\alias{n_sync}
\title{NSYNC}
\usage{
n_sync(justin = NULL, jc = NULL, chris = NULL, joey = NULL, lance = NULL)
}
\arguments{
\item{justin}{A value to compare.}

\item{jc}{A value to compare.}

\item{chris}{A value to compare.}

\item{joey}{A value to compare.}

\item{lance}{A value to compare.}
}
\description{
Are they actually still i\strong{n-sync}?
}
