% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock.R
\name{phil_columns}
\alias{phil_columns}
\title{Phil columns}
\usage{
phil_columns(.data, value, play = FALSE)
}
\arguments{
\item{.data}{A data frame to add a new column to.}

\item{value}{The value to fill the new column with.}

\item{play}{Logical, whether to play the track on Spotify.}
}
\description{
Phil a column!
}
