% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify.R
\name{play_music}
\alias{play_music}
\title{Play music from R}
\usage{
play_music(track)
}
\arguments{
\item{track}{A Spotify track in this format: \code{"spotify:track:6mYXXXgFZZmk2mgejwSwdN"}}
}
\description{
Start to play a song on your local Spotify from the R console. Standalone this is pretty useless, but makes total sense in \strong{musicplyr}...
}
