% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock.R
\name{slizz_top}
\alias{slizz_top}
\title{Slizz top}
\usage{
slizz_top(.data, ..., play = FALSE)
}
\arguments{
\item{.data}{A data frame to be slizzed.}

\item{...}{Additional arguments passed to \code{\link[dplyr]{slice_head}()}.}

\item{play}{Logical, whether to play the track on Spotify.}
}
\description{
Slizze the top of your data frame.
}
