% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninetys.R
\name{fill_collins}
\alias{fill_collins}
\title{Fill Collins}
\usage{
fill_collins(.data, value, play = FALSE)
}
\arguments{
\item{.data}{A data frame to add a new column to.}

\item{value}{The value to fill the new column with.}

\item{play}{Logical, whether to play the track on Spotify.}
}
\description{
Fill a collin!
}
