% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock.R
\name{phil_columns}
\alias{phil_columns}
\title{Phil columns}
\usage{
phil_columns(.data, ...)
}
\arguments{
\item{.data}{A data frame to be modified.}

\item{...}{Additional arguments to be passed to the mutate function.}
}
\description{
Phil a column!
}
